Fakta taulujen FLOAT muutettu FLOAT(10,3)

+++ sql/cooldev_olap.sql
@@ -51,7 +51,7 @@ ENGINE = InnoDB;
 CREATE TABLE IF NOT EXISTS `cooldev_olap`.`measurements_fact` (
   `sensor_key` INT NOT NULL,
   `date_key` INT NOT NULL,
-  `value` FLOAT NOT NULL,
+  `value` FLOAT(10,3) NOT NULL,
   INDEX `fk_measurements_fact_sensors_dim_idx` (`sensor_key` ASC),
   PRIMARY KEY (`sensor_key`, `date_key`),
   INDEX `fk_measurements_fact_dates_dim1_idx` (`date_key` ASC),
@@ -74,7 +74,7 @@ ENGINE = InnoDB;
 CREATE TABLE IF NOT EXISTS `cooldev_olap`.`total_consumptions_fact` (
   `sensor_key` INT NOT NULL,
   `date_key` INT NOT NULL,
-  `value` FLOAT NOT NULL,
+  `value` FLOAT(10,3) NOT NULL,
   PRIMARY KEY (`sensor_key`, `date_key`),
   INDEX `fk_total_consumptions_fact_dates_dim1_idx` (`date_key` ASC),
   CONSTRAINT `fk_total_consumptions_fact_sensors_dim1`
@@ -96,7 +96,7 @@ ENGINE = InnoDB;
 CREATE TABLE IF NOT EXISTS `cooldev_olap`.`productions_fact` (
   `sensor_key` INT NOT NULL,
   `date_key` INT NOT NULL,
-  `value` FLOAT NOT NULL,
+  `value` FLOAT(10,3) NOT NULL,
   PRIMARY KEY (`sensor_key`, `date_key`),
   INDEX `fk_productions_fact_dates_dim1_idx` (`date_key` ASC),
   CONSTRAINT `fk_productions_fact_sensors_dim1`
@@ -118,7 +118,7 @@ ENGINE = InnoDB;
 CREATE TABLE IF NOT EXISTS `cooldev_olap`.`heating_consumptions_fact` (
   `sensor_key` INT NOT NULL,
   `date_key` INT NOT NULL,
-  `value` FLOAT NOT NULL,
+  `value` FLOAT(10,3) NOT NULL,
   PRIMARY KEY (`sensor_key`, `date_key`),
   INDEX `fk_heating_consumptions_fact_dates_dim1_idx` (`date_key` ASC),
   CONSTRAINT `fk_heating_consumptions_fact_sensors_dim1`
@@ -140,7 +140,7 @@ ENGINE = InnoDB;
 CREATE TABLE IF NOT EXISTS `cooldev_olap`.`lighting_consumptions_fact` (
   `sensor_key` INT NOT NULL,
   `date_key` INT NOT NULL,
-  `value` FLOAT NOT NULL,
+  `value` FLOAT(10,3) NOT NULL,
   PRIMARY KEY (`sensor_key`, `date_key`),
   INDEX `fk_lights_consumptions_fact_dates_dim1_idx` (`date_key` ASC),
   CONSTRAINT `fk_lights_consumptions_fact_sensors_dim1`
@@ -162,7 +162,7 @@ ENGINE = InnoDB;
 CREATE TABLE IF NOT EXISTS `cooldev_olap`.`outlets_consumptions_fact` (
   `sensor_key` INT NOT NULL,
   `date_key` INT NOT NULL,
-  `value` FLOAT NOT NULL,
+  `value` FLOAT(10,3) NOT NULL,
   PRIMARY KEY (`sensor_key`, `date_key`),
   INDEX `fk_outlets_consumption_fact_dates_dim1_idx` (`date_key` ASC),
   CONSTRAINT `fk_outlets_consumption_fact_sensors_dim1`
@@ -181,3 +181,5 @@ ENGINE = InnoDB;
 SET SQL_MODE=@OLD_SQL_MODE;
 SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS;
 SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS;
+
+
